check if .tox file exists, before trying to copy it

+++ src/tox_client.cpp
@@ -246,11 +246,13 @@ void ToxClient::saveToxProfile(void) {
 		return;
 	}
 
-	std::filesystem::copy_file(
-		_tox_profile_path,
-		_tox_profile_path + ".old",
-		std::filesystem::copy_options::overwrite_existing
-	);
+	if (std::filesystem::exists(_tox_profile_path)) {
+		std::filesystem::copy_file(
+			_tox_profile_path,
+			_tox_profile_path + ".old",
+			std::filesystem::copy_options::overwrite_existing
+		);
+	}
 
 	std::filesystem::rename(
 		tmp_path,
